* kdc_preauth.c (get_etype_info): Initialize length of 	etype_info_entry to KRB5_ETYPE_NO_SALT. Before copying salt into 	etype_info_entry, test that salt length (from get_salt_from_key()) 	is not SALT_TYPE_NO_LENGTH.
	* kdc_util.c (get_salt_from_key): Initialize salt type to
	SALT_TYPE_NO_LENGTH (instead of -1).

	* network.c (process_packet): Cast length in sendto() call to int.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12787 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/kdc/network.c
@@ -468,7 +468,7 @@ void process_packet(port_fd, prog, portnum)
 	com_err(prog, retval, "while dispatching");
 	return;
     }
-    cc = sendto(port_fd, response->data, response->length, 0,
+    cc = sendto(port_fd, response->data, (int) response->length, 0,
 		(struct sockaddr *)&saddr, saddr_len);
     if (cc == -1) {
 	char addrbuf[46];
